requirements.txt
flask==3.1.2
requests==2.32.5
mysql-connector-python==9.4.0